Disable some pylint complaints
I am not in general a fan of slapping on 'pylint: disable' when pylint
complains... however I'm not really sure how to fix it in a better way.
Also, there is precedent in this project for ignoring either of these
tests (too-many-arguments, too-many-instance-attributes).  I'm happy to
handle this in a better way if desired.